adding rating

+++ src/main/java/layer/data/api/RatingDataService.java
@@ -1,11 +1,9 @@
 package layer.data.api;
 
-import java.math.BigDecimal;
-
 public interface RatingDataService {
 
   RatingData rate(Long idMovie);
 
-  void rate(Long idUser, Long idMovie, BigDecimal value) throws DataException;
+  void rate(Long idUser, Long idMovie, int value) throws DataException;
 
 }
